+++ mds/apis/prv_api/service_areas.py
@@ -14,10 +14,7 @@ from mds.access_control.scopes import SCOPE_PRV_API
 from mds.apis import utils
 
 from rest_framework import filters
-<<<<<<< HEAD
 
-=======
->>>>>>> added alphabetical order for Polygon and Areas ViewSet
 
 class PolygonRequestSerializer(serializers.ModelSerializer):
     """What we expect for a geographic polygon.
@@ -79,13 +76,8 @@ class PolygonViewSet(utils.MultiSerializerViewSetMixin, viewsets.ModelViewSet):
     permission_classes = (require_scopes(SCOPE_PRV_API),)
     queryset = models.Polygon.objects.prefetch_related("areas").all()
     filter_backends = (filters.OrderingFilter,)
-<<<<<<< HEAD
     ordering_fields = ("label",)
     ordering = "label"
-=======
-    ordering_fields = ('label',)
-    ordering = ('label')
->>>>>>> added alphabetical order for Polygon and Areas ViewSet
     lookup_field = "id"
     serializer_class = PolygonResponseSerializer
     serializers_mapping = {
@@ -202,8 +194,8 @@ class AreaViewSet(utils.MultiSerializerViewSetMixin, viewsets.ModelViewSet):
     permission_classes = (require_scopes(SCOPE_PRV_API),)
     queryset = models.Area.objects.prefetch_related("polygons").all()
     filter_backends = (filters.OrderingFilter,)
-    ordering_fields = ('label',)
-    ordering = ('label')
+    ordering_fields = ("label",)
+    ordering = "label"
     lookup_field = "id"
     serializer_class = AreaResponseSerializer
     serializers_mapping = {